chore(ci): remove concurrency on push main
When the push-main workflow finishes it pushes to main, triggering another run of the push-main workflow. That one will be immediately cancelled because it's a version commit. The previous workflow will however be cancelled, which is not what we want.

Concurrency doesn't allow dynamic values, so a dynamic build step would have to be added to avoid 2 PR merges trying to up the version